Aula_05/requirements.txt
django
djangorestframework
django-cors-headers
pillow
setuptools